+++ src/main/java/roomescape/controller/ReservationController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 import roomescape.service.ReservationService;
-import roomescape.controller.dto.ReservationRegisterDto;
-import roomescape.controller.dto.ReservationResponseDto;
+import roomescape.service.dto.ReservationRegisterDto;
+import roomescape.service.dto.ReservationResponseDto;
 
 @RestController
 @RequestMapping("/reservations")
