+++ src/main/java/roomescape/service/ReservationService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-import roomescape.controller.dto.ReservationRegisterDto;
-import roomescape.controller.dto.ReservationResponseDto;
+import roomescape.service.dto.ReservationRegisterDto;
+import roomescape.service.dto.ReservationResponseDto;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationRepository;
